Github: refactor module
